Update README.md
Gender and Ethnic Group Analysis:

We investigated the impact of gender and ethnic group differences on exam performance. This analysis guides us in matters of equality and justice in education.

The Role of Family and Social Status:

We examined factors such as parental status, marital status, and the number of siblings to understand their impact on student achievements. This helped emphasize the importance of family and social support factors in education.

Lunch Type and Exam Preparation:

We analyzed the relationship between lunch type and exam preparation. This helps us understand the influence of students' daily lives on their exam performance.